Externalize steps in frozen files erigondb setting (#17448)
This is part 2/n of https://github.com/erigontech/erigon/issues/16765 .
Not essential for it, but it removes the need to recompile the code
every time the DB geometry is changed.

#### Overview

The main use is when, let's say, you make step size -> step size / 2,
then in frozen files you need to make step in frozen files -> step in
frozen files * 2, i.e., using E32 default settings that would mean
changing the limit 64 -> 128.

Another possible use would be for experimenting with bigger merge limits
(without changing step size), we could use this param to relax the
default limit of 64 steps.

BTW, in fact I tested this one just for curiosity on an existing mainnet
node, in fact it created some 0-128 files, but it didn't create 128-256
immediately, nor it cleaned 0-64, etc.

probably it has to do with how we apply merging + how the current
"pyramid merging" is coded, if someone wants to experiment with bigger
merge sizes those will need to be handled separately (out of scope of
this PR).

#### Implementation details

The stepsInFrozenFile param is exposes at Aggregator level, and
propagates down to every dependency who needs it, down to the merging
process.

The existing constant `StepsInFrozenFile` is renamed to
`DefaultStepsInFrozenFile` and retrofitted as the default value injected
into Aggregator.

CLI params to `erigon` and `rpcdaemon` are exposed as ways to override
this value.

A loud warning is logged at start because those are not common params to
be overriden, i.e., it is expected that DB geometry will be handled by
tools, but since we don't have [yet] enough tools to manipulate and
introspect the DB, we assume such transformations are done manually by
erigon devs and params are not persisted and feed into erigon by
scripts.