Fix validators producing bad blocks on Hoodi (#17487)
This fixes the issue we've had with proposing bad blocks on hoodi.

The majority of the changes here are tracing which I have ported from
main. The underlying cause of the issue turned out to be that for mining
we where not initializing block execution. Which effectively means not
calling system contracts. This is ok so long as there are no side
effects which are need by transactins in the block.

However we're now calling the:

```
"BEACON_ROOTS_ADDRESS": "0x000f3df6d732807ef1319fb7b8bb8522d0beac02",
```

And if a transaction needs to see the result of this and performs
processing based on it they will fail.

The fix is just to add:

```
core.InitializeBlockExecution(cfg.engine, chainReader, current.Header, cfg.chainConfig, ibs, &state.NoopWriter{}, logger, nil)
```

before we start processing blocks.

(cherry picked from commit 86ad509bed8a59be485e624d49da1d5e794c5a20)