rpcdaemon: fix Hive test eth_getTransactionReceipt pre-byzantium (#17509)
This PR fixes the following Hive RPC compatibility tests:
- eth_getTransactionReceipt/get-legacy-contract (erigon_default)
- eth_getTransactionReceipt/get-legacy-input (erigon_default)
- eth_getTransactionReceipt/get-legacy-receipt (erigon_default)

For pre-Byzantium blocks, the Ethereum specification requires that
transaction receipts contains:
    the "postState" field and
    "NOT" the status field.

To calculate "postState" it is necessary enable the commitment
The current database and snapshots for pre-Byzantium blocks do not
contain valid postState data, therefore, we must explicitly re-calculate
these receipts to be compliant to the specification (postState instead
of status).

If "commitment is not enabled", we always provide the "status" field,
regardless of the block number.

Some tests has been implemented to see the fixes using the hive DB

In order to see these tests successfully solved in Hive, a PR enabling
commitment is required in the Hive repository itself