Refactor txNum ranges -> step ranges logic (#17535)
`FilesItem.startTxNum/endTxNum` -> `from/toStep` logic is used in quite
a bunch of places in the code.

Although the math is simple, it makes sense to push such behavior into
FilesItem itself. And the result can be embedded into a function call if
it expects exactly (from, to), which is most of the places, making the
code cleaner.

Also there is a trick in how start/endTxNum is calculated at struct
initialization (and unchecked) that if not respected may lead to
unexpected results; I added some docs explaining it.