eth_getProof: Ignore root hash check in GenerateWitness (#17551)
This fixes the following issues:

- example 2 in https://github.com/erigontech/erigon/issues/16752
(therefore allowing us to close the issue)
- https://github.com/erigontech/erigon/issues/17454

Probably also fixes https://github.com/erigontech/erigon/issues/17488 if
the empty branchdata problem in the files is overcome, as I deliberately
crafted this examples to trigger the same scenario, as explained below.

## Explanation 
All the failing examples above represent the same category :
_proofs for non-existent accounts, where the last node in the merkle
proof is an extension node with a diverging key/path._
 
 
Now, the section in the eth_getProof EIP says if the last matching node
is a branch node then the proof value in the node should be an empty
one:
https://eips.ethereum.org/EIPS/eip-1186#proofs-for-non-existent-values .
But it doesn't specify what should happen if the last node is an
extension node.
 
Looking at the proofs returned by Geth, and looking at the verification
logic an how it applies to this scenario, in `func verifyProof(root
common.Hash, key []byte, proofs map[common.Hash]Node, used
map[common.Hash]rawProofElement) ([]byte, error) {` the relevant code is
:
 
 ```Go
 	case *ShortNode:
			shortHex := nt.Key
			if len(shortHex) > len(key) {
return nil, fmt.Errorf("len(shortHex)=%d must be leq len(key)=%d",
len(shortHex), len(key))
			}
			if !bytes.Equal(shortHex, key[:len(shortHex)]) {
				return nil, nil
			}
			node, key = nt.Val, key[len(shortHex):]
```

So, when the last node in the proof is an extension node (aka `ShortNode`), and it's key doesn't match with the hashedKey in the requested account, then the proof verifies. What's noteworthy here is that the value (i.e. pointer to the next node) is irrelevant and can be ignored.

In Erigon, when this scenario happens, the last node in the proof is `ShortNode{ Key: divergingKey, Val: nil}` . 

This means that the proof verification succeeds, but the root hash of the proof trie doesn't match state root of that block due to `Val: nil`, so an error is thrown before the proof verification can run.

We can safely skip this root hash check however, since the proof will still verify. 

In the future, to maintain consistency in the root hash, we can replace the `Val: nil` by `Val: HashNode(hash)` , where `hash` is the `hash` of the node the extension node points to. This would require extra work to load the next node (somewhere outside the grid), and collapse it to a `HashNode`. For the time being, this PR should suffice.

---------

Co-authored-by: antonis19 <antonis19@users.noreply.github.com>