Dump tool to print history snapshots content (#17864)
This PR is related to the discussion in this task (comments:
https://github.com/erigontech/erigon/issues/17370#issuecomment-3514260624
https://github.com/erigontech/erigon/issues/17370#issuecomment-3514571128)

Specifically, it concerns the issue I raised about history files
containing a large number of unnecessary duplications that we can safely
eliminate.

The core of the problem is that the Elias–Fano sequences in `.ef` files
for different keys are longer than they need to be — they sometimes
include transaction numbers even when no actual data change occurred
(the value remained the same).
This PR implements a helper command that makes it possible to detect
such cases.

Regarding the `HistoryRange` method — It doesn’t show a stream of
changes; instead, it returns only a single point for each key that
changed. The function’s own description in the code reads:
```
// HistoryRange producing state-patch for Unwind - return state-patch for Unwind:
// "what keys changed between `[from, to)` and what was their value BEFORE txNum"
```

As a result, we can compress history files — not only `.v` but also
`.ef` and accessor files — several times over. For Gnosis, the amount of
redundant duplication is roughly **2×**.