switch RequestMore to use channels (#17881)
This potentially fixes:
https://github.com/erigontech/erigon/issues/17590

Or at leaste it does not make it worse. From the goroutine trace it
seemed like the timers here may be deadlocked. This is based on the fact
that they are all asleep in the dump, and looking at the logs at leaste
one of them ought to have fired. As this is a guess and not a definitae
fix I have run qa many times in repetition just to make sure it doesnt
make the situation worse.

I think it also makes the code slightly more idiomatic go.