BAL Parallel Exec Integration (#17804)
This PR contains the initial implementation between the parallel
execution code and a block's BAL.

The in the implemented code if a block contains a BAL the version map
for the block executer will be pre populated and then all transactions
will be optimistically scheduled with no dependency resolution.

Still to be done is the outbound BAL creation & testing. This should be
done here:


[erigon/execution/stagedsync/exec3_parallel.go](https://github.com/erigontech/erigon/blob/d069a222a726a96ec42721945bfeab151fee814e/execution/stagedsync/exec3_parallel.go#L210)

Line 210 in
[d069a22](https://github.com/erigontech/erigon/commit/d069a222a726a96ec42721945bfeab151fee814e)

 // TODO --- BAL Implementation integration point --- 
Please read the notes there to get an idea on how to proceed.

To enable parallel exec set the following environment variable:

export ERIGON_EXEC3_PARALLEL=true
and optionally:

export ERIGON_EXEC3_WORKERS=num workers
Note: Testing on a live chain will likely need
https://github.com/erigontech/erigon/pull/17783 merged first