Added benchmark test for random access history state of accounts (#18058)
This test iterates over N random blocks in history to select accounts,
which are then used for eth_getBalance RPC calls.
```
./build/bin/rpctest benchEthGetBalanceRandomAccount --randBlocks=10000
```

For Sepolia (1.7 TB for the archive node and 82 GB for account history
snapshots), it takes about **0.6–0.8 ms** per account on average. The
fluctuations in average access speed (20-25% range) for the test are
quite large, even when a large number of blocks are sampled with
`--randBlocks`.

The access speed is influenced by the size of the snapshot files — for
example, in the case of deduplication (which reduces the size by almost
1.5×), the average access time improves to about **0.5–0.6** ms per
account.

Page-level compression (set to 16) did not produce any change in access
speed, even though the snapshots became about 1.5x as small.