Revert "Remove duplicates from AccHistory and StorageHistory" (#18109)
Reverts erigontech/erigon#17962

<@525860938360815626> : Mike prepared a DB with commitments enabled for
historical eth_getProof() verifications.
Upon executing the historical tests, we observed that the state root
calculation failed for few tests with postState and so verification of
receiptsRoot(the calculated receipt root doesn't match
block.receiptRoot).
Previously, even when running the Hive tests (with commitment enabled),
the state root calculation was correct.
By bisecting the commits, we identified that the failure was introduced
after the merge of PR #17962.
PR #18108 enables historical tests with Erigon commitments and inserts
two calls to TouchKey() in DomainPut(). This addition resolves the
failed state root calculation and the verification of the receipts root.
Since I am not an expert in this specific code component, could you
please verify if this is the correct fix?