Final downloader cleanup (#18056)
Fixes a few things that bothered me that weren't right in 3.1. #16532,
#17664, #17310, #16654.

#16698 is not addressed here (not sure if I will, but testing this might
make it worthwhile).

List of changes:

- Downloader.Add is split into Seed and Download. This prevents allowing
downloads to occur in parts of code that expect to only Seed (like in
the pruning code). It also prevents incomplete/partial torrents being
added. Download now blocks and logs synchronously until all requested
snapshots are complete. This simplifies the API back and forth with the
caller. Snapshot requests now include an appropriate name for the
request set.
- Download.Completed, SetLogPrefix are no longer required. These don't
work well since Downloader can be shared, and those concepts don't
support sharing.
- A few logs that weren't clear they weren't related to the
downloader/BitTorrent were adjusted.
- A bunch of metrics in webseed requesting that aren't needed anymore
were removed.
- Download.AddTorrentsFromDisk can be called without breaking existing
downloads. It only adds completed snapshots, it's assumed anything else
would be a mistake. It's no longer the default anywhere.
- torrent.TorrentSpec is not used directly anymore. It's been deprecated
for a long time.
- I fixed STOP_BEFORE_STAGE and STOP_AFTER_STAGE.
- There's no longer snapshot "validation" on completion. We're trusting
the torrent client to do the right thing, but also not asking it to
download things that should already be complete.
- Data is always rechecked if an infohash changes.
- Torrent/metainfo files are retained on disk before initial sync
completes. This means significantly faster start-up times if a sync is
resumed midway.