Added dedup mechanism in history snapshots (#17958)
This current PR modifies the history snapshot merge mechanism (including
indexes) so that redundant values do not end up in the output file.

This provides roughly a 2× reduction in disk usage for both `.v` and
`.vi` files, as well as `.ef`.

Frozen files (those 64 steps in size) will require regeneration or an
increased `mergeLimit`, since they cannot be processed in the
background.

In any case, there will be an additional PR adding a tooling command to
rebuild history snapshots, and there may also be a PR to increase
mergeLimit (issue https://github.com/erigontech/erigon/issues/17369).

---------

Co-authored-by: Alexey Sharov <AskAlexSharov@gmail.com>