Remove extra --chain flag to snapshots reset
It works for datadir because of a specially implement flag type, but chain does not use the same thing. As a result, if you pass --chain before snapshots reset, it is silently ignored and you risk resetting to the wrong chain.

urfave/cli@v3 added interspersed flags for this, but we are on v2. I previously made a change to make it work as expected for datadir but didn't test chain.