Pull torrent excessive wantPeers event fix (#18256)
This fixes excessive updates in the announcer work queue due to an
excessively triggered event in the torrent client
[here](https://github.com/anacrolix/torrent/commit/f946a94ecebc77fa7fd0a471b37738f4d35676da#diff-449c382592420ab0e94eb0707f863e9fdaeceea483a1f39ebce93c42b57ba04dL52-L1726).

This caused extremely high CPU when torrents exceeded around 13k in
count and caused a bottleneck on the client it couldn't recover from. It
didn't stop anything from operating it just got very slow, without any
scheduler issues or thrashing.

There's still more aggressive improvements on main but I tried to keep
it minimal.

The behaviour matches the profile in
https://github.com/erigontech/erigon/issues/18162 so it should be an
exact fix for that. I think it also finally fixes
https://github.com/erigontech/erigon-qa/issues/320.