cl/beacon: add single_attestation event topic support (#18142)
This PR fixes the Beacon API events endpoint to publish
SingleAttestation objects (introduced in Electra) under the
`single_attestation` topic instead of `attestation`.

Currently, Caplin publishes all attestations under the `attestation`
topic, regardless of whether they are pre-Electra `Attestation` objects
or post-Electra `SingleAttestation` objects.

**References:**
-
https://github.com/ethereum/consensus-specs/blob/dev/specs/electra/p2p-interface.md#beacon_attestation_subnet_id
-
https://github.com/ethereum/beacon-APIs/blob/master/apis/eventstream/index.yaml#L71
-
https://github.com/ethereum/consensus-specs/blob/dev/specs/electra/beacon-chain.md#singleattestation

**For context:**

We'd love to see this added. We have some contributoors pushing to the
ethPandaOps Xatu dataset using Caplin - but we're not capturing
attestations because of it. **For example:**
https://lab.ethpandaops.io/xatu/contributors/abstractdrank74#attestation-latency