data files: add version into meta-header (#18225)
This PR changes the internal format of snapshot files, specifically how
the files are written to disk:

1. The logic of the first byte in the file has been redefined — it is
now used for snapshot versioning. Before this PR, this byte was always
0x00, which effectively meant version 0. In this case, the compression
and decompression behavior remains unchanged.
2. For version 0x01, two additional bytes are added at the beginning of
the file:
the first indicates the version, the second is a feature bitmask used
during compression of the file.
3. At the moment, three bits are reserved: page-level compression, keys
compression, and values compression.
If the page-level compression flag is enabled, a third additional byte
follows the first two bytes in the file. It indicates the compression
level — the number of elements per page.
4. The rest of the file has the same format as in version 0.